Отправляет на проверку 4дз
